backend/pyproject.toml
[project]
name = "studying"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.1",
    "msgspec>=0.20.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.11.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = []
